修改sql

+++ jsd-manager/src/main/java/com/ald/jsd/mgr/biz/service/MgrBorrowCashRenewalService.java
@@ -1,12 +1,14 @@
 package com.ald.jsd.mgr.biz.service;
 
 
+import org.apache.ibatis.annotations.Param;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public interface MgrBorrowCashRenewalService {
 
-    Integer getRenewalAmountByDays( Date startDate,Date endDate);
+    BigDecimal getRenewalAmountByDays( Date startDate,Date endDate);
 
     int getRenewalPersonsByDays(Integer days);
